requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.0,<6.0
drf-spectacular>=0.26,<1.0
openpyxl>=3.1,<4.0
Pillow>=10.0,<11.0